1.3.0 - refactor support for signature and hash function
